requirements.txt
django==4.1.2
djangorestframework==3.14.0
sqlparse==0.4.3
psycopg2-binary==2.9.4
pytest-django==4.5.2
pytest==7.1.3